requirements.txt
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
langchain-google-genai==2.0.10
flask
python-dotenv
langchain==0.3.27
duckduckgo-search
openai
gunicorn